\name{PBIBD-package}
\alias{PBIBD-package}
\alias{PBIBD}
\docType{package}
\title{Partially Balanced Incomplete Block Designs}
\description{Assists in calculating the efficiencies of Partially Balanced Incomplete Block Designs with any number of associate classes. This will help the researchers in calculating the efficiencies of their PBIB designs very quickly and efficiently.}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PBIBD}
\packageIndices{PBIBD}
}
\author{
Parneet Kaur <parneet.nonu93@gmail.com>

Kush Sharma <kush.vashishtha@gmail.com>

Davinder Kumar Garg <dkgarg_stat@yahoo.co.in>

Maintainer: Kush Sharma <kush.vashishtha@gmail.com>
}
\note{This package is currently under intensive development and changes are to be expected in the near future.}
\references{
Dey,A. (1986). Theory of block designs.Wiley Eastern Limited, New Delhi

Garg, D.K. and Mishra, G.(2013). General solution of normal equations in the intra-block analysis of PBIB designs with four associates. Journal of Mathematics and System Sciences. 3,412-416

Garg, D.K., and Singh, G.P. (2014). Analysis and general solution of normal equations for five associate class PBIB designs. Global Journal for Research Analysis, 3(11), 16-19

Garg, D.K., and Singh, G.P. (2015). General solution of normal equations in the intra-block analysis of PBIB designs with any (m>=2) number of associate classes. American Journal of Sustainable Cities and Society, 4(11), 196-202

Rao, C. R. (1947b). General methods of analysis for incomplete block designs, J. Amer. Statist. Assoc. 42, 541-561
}
\keyword{Partially Balanced Incomplete Block Designs}
\keyword{Efficiencies}
\keyword{Two-associate class PBIB Design}
\keyword{Three-associate class PBIB Design}
\keyword{Four-associate class PBIB Design}
\keyword{Five-associate class PBIB Design}
\keyword{m-associate class PBIB Design}

