\name{PBIBD-package}
\alias{PBIBD-package}
\alias{PBIBD}
\docType{package}
\title{Partially Balanced Incomplete Block Designs}
\description{The PBIB designs are important type of incomplete block designs having wide area of their applications for example in agricultural experiments, in plant breeding, in sample surveys etc. This package constructs various series of PBIB designs and assists in checking all the necessary conditions of PBIB designs and the association scheme on which these designs are based on. It also assists in calculating the efficiencies of PBIB designs with any number of associate classes. The package also constructs Youden-m square designs which are Row-Column designs for the two-way elimination of heterogeneity. The incomplete columns of these Youden-m square designs constitute PBIB designs. With the present functionality, the package will be of immense importance for the researchers as it will help them to construct PBIB designs, to check if their PBIB designs and association scheme satisfy various necessary conditions for the existence, to calculate the efficiencies of PBIB designs based on any association scheme and to construct Youden-m square designs for the two-way elimination of heterogeneity.}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PBIBD}
\packageIndices{PBIBD}
}
\author{Parneet Kaur <parneet.nonu93@gmail.com>

Kush Sharma <kush.vashishtha@gmail.com>

Davinder Kumar Garg <dkgarg_stat@yahoo.co.in>

Maintainer: Kush Sharma <kush.vashishtha@gmail.com>
}
\note{This package is currently under intensive development and changes are to be expected in the near future.}
\references{
Dey, A. (1986). Theory of block designs.Wiley Eastern Limited, New Delhi

Garg, D.K., and Singh, G.P. (2015). General solution of normal equations in the intra-block analysis of PBIB designs with any (m>=2) number of associate classes. American Journal of Sustainable Cities and Society, 4(11), 196-202

Kaur, P. and Garg, D. K. (2016). Construction of some higher associate class PBIB designs using symmetrically repeated differences, Arya Bhatta Journal of Mathematics and Informatics, 8(2), 65-78.

Rao, C. R. (1947b). General methods of analysis for incomplete block designs, J. Amer. Statist. Assoc. 42, 541-561

Sharma, K. and Garg, D. K. (2017) Construction of Three associate PBIB designs using some sets of initial blocks. International Journal of Agricultural and Statistical Sciences (IJASS), 13(1): 55-60 

Sharma, K. and Garg, D. K. (2017). m-associate PBIB designs using Youden-m Squares. Communications in Statistics-Theory and Methods, In press. DOI: 10.1080/03610926.2017.1324990
}
\keyword{Partially Balanced Incomplete Block Designs}
\keyword{Efficiencies}
\keyword{Necessary conditions}
\keyword{m-associate class PBIB Design}
\keyword{Construction of PBIB designs}
\keyword{Symmetric differences}
\keyword{Youden-m squares}
\keyword{Youden-m square designs}
\keyword{Row-Column designs}
\keyword{Two-way elimination of heterogenity}
