\name{PBIBD-package}
\alias{PBIBD-package}
\alias{PBIBD}
\docType{package}
\title{Efficiencies of PBIB Designs}
\description{Assists in calculating the efficiencies of any type of Partially Balanced Incomplete Block Designs with two, three, four and five associate classes. This will help the researchers in calculating the efficiencies of their PBIB designs very quickly and efficiently.}
\details{
\tabular{ll}{
Package: \tab PBIBD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-08-25\cr
License: \tab GPL(>=2)\cr
}
}
\author{
Parneet Kaur <parneet.nonu93@gmail.com>

Kush Sharma <kush.vashishtha@gmail.com>

Davinder Kumar Garg <dkgarg_stat@yahoo.co.in>

Maintainer: Kush Sharma <kush.vashishtha@gmail.com>
}
\note{This package is currently under intensive development and changes are to be expected in the near future.}
\references{
Dey,A. (1986). Theory of block designs.Wiley Eastern Limited, New Delhi

Garg, D.K. and Mishra, G.(2013). General solution of normal equations in the intra-block analysis of PBIB designs with four associates. Journal of Mathematics and System Sciences. 3,412-416

Garg, D.K., and Singh, G.P. (2014). Analysis and general solution of normal equations for five associate class PBIB designs. Global Journal for Research Analysis, 3(11), 16-19

Rao, C. R. (1947b). General methods of analysis for incomplete block designs, J. Amer. Statist. Assoc. 42, 541-561
}
\keyword{Partially Balanced Incomplete Block Designs}
\keyword{Efficiencies}
\keyword{Two-associate class PBIB Design}
\keyword{Three-associate class PBIB Design}
\keyword{Four-associate class PBIB Design}
\keyword{Five-associate class PBIB Design}

